Add observer to slot
The observer will trigger `_moveCaret` every time it changes. This is
mainly because if the navigation items changes (logged out then logged
in user) it wouldn't recalculate the caret position. This way it will do
it.

The issue is that for some reason I can't understand, there is a delay
between the target being available on the DOM, not being able to
calculate properly the position. I have no idea why but if you just
access the target's offset, it will work as expected.