Adjust caret position
Take the width of the caret into account when positioning it in the center of the nav item to allow this to be properly centered when the nav item doesn't have a label, but an icon only.

Also removes the margin on the last nav item.

Part of the changes for this Trello card: https://trello.com/c/Duf5kL2m